memcached parseTTL return value changed int to int32

+++ state/memcached/memcached.go
@@ -20,8 +20,8 @@ const (
 	timeout            			= "timeout"
 	ttlInSeconds       			= "ttlInSeconds"
 	// These defaults are already provided by gomemcache
-	defaultMaxIdleConnections 	= 2
-	defaultTimeout            	= 1000 * time.Millisecond
+	defaultMaxIdleConnections	= 2
+	defaultTimeout				= 1000 * time.Millisecond
 )
 
 type Memcached struct {
@@ -103,7 +103,7 @@ func getMemcachedMetadata(metadata state.Metadata) (*memcachedMetadata, error) {
 	return &meta, nil
 }
 
-func (r *Memcached) parseTTL(req *state.SetRequest) (int, error) {
+func (r *Memcached) parseTTL(req *state.SetRequest) (int32, error) {
 	if val, ok := req.Metadata[ttlInSeconds]; ok && val != "" {
 		parsedVal, err := strconv.ParseInt(val, 10, 0)
 		if err != nil {
@@ -114,7 +114,7 @@ func (r *Memcached) parseTTL(req *state.SetRequest) (int, error) {
 			return 0, nil
 		}
 
-		return int(parsedVal), nil
+		return int32(parsedVal), nil
 	}
 
 	return 0, nil
@@ -128,7 +128,7 @@ func (m *Memcached) setValue(req *state.SetRequest) error {
 	}
 
 	bt, _ = utils.Marshal(req.Value, m.json.Marshal)
-	err = m.client.Set(&memcache.Item{Key: req.Key, Value: bt, Expiration: int32(ttl)})
+	err = m.client.Set(&memcache.Item{Key: req.Key, Value: bt, Expiration: ttl})
 	if err != nil {
 		return fmt.Errorf("failed to set key %s: %s", req.Key, err)
 	}
